Add MessageArea.js
